Add course view page with delete and edit function
Created new course view page that shows once user clicks on a course on
the main page. This fragment shows all the course details and has an
edit button and delete FAB.

Created new EditCourseFragment which is similar to adding a new course,
but it has prepopulated fields and immuatable course ID. On completion,
brings user back to course view page.

Added delete functionality, and on completion brings user to home page.